String programs added

+++ src/com/Lokesh/LokeshArrays/FactorsOfANumber.java
@@ -5,7 +5,7 @@ import java.util.*;
 public class FactorsOfANumber {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
-        int number = s.nextInt();
+        long number = s.nextInt();
         for(int i = 1; i<=number; i++){
             if(number%i == 0 ){
                 System.out.println(i + " ");
